Change уÑexception handling (add MethodNotAllowedException, HttpBadRequestException)

+++ app/Action/TaskDeleteAction.php
@@ -8,6 +8,7 @@ use Laminas\Diactoros\Response\JsonResponse;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Server\RequestHandlerInterface;
+use Slim\Exception\HttpBadRequestException;
 
 class TaskDeleteAction implements RequestHandlerInterface
 {
@@ -21,6 +22,10 @@ class TaskDeleteAction implements RequestHandlerInterface
     public function handle(ServerRequestInterface $request): ResponseInterface
     {
         $params = $request->getParsedBody();
+        $id = $params['id'];
+        if (!$id) {
+           throw new HttpBadRequestException($request);
+        }
         $this->taskRepo->delete($params['id']);
         return (new EmptyResponse())->withHeader('Location', "/tasks");
     }
